% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{AddMan}
\alias{AddMan}
\title{Expand a Landscape object by copying an existing management scenario}
\usage{
AddMan(obj, tocopy = 1)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{tocopy}{positive integer greater than 1, the number of replicate management options}
}
\description{
Expand a Landscape object by copying an existing management scenario
}
\author{
T. Carruthers
}
