% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{calcCB}
\alias{calcCB}
\title{Calculate stocking cost and benefit (effort) for a resolved landscape object}
\usage{
calcCB(obj)
}
\arguments{
\item{obj}{an object of class 'landscape' for which an IFD calculation using calceff() has been carried out.}
}
\description{
Calculate stocking cost and benefit (effort) for a resolved landscape object
}
\author{
T. Carruthers
}
