% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{calccouch2}
\alias{calccouch2}
\title{Calculate couch effect (the attractivity of other activities given a fitted landscape object)}
\usage{
calccouch2(obj)
}
\arguments{
\item{obj}{an object of class 'Landscape'}
}
\description{
Calculate couch effect (the attractivity of other activities given a fitted landscape object)
}
\author{
T. Carruthers
}
