% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calceff.R
\name{calceff}
\alias{calceff}
\title{The IFD effort calculator}
\usage{
calceff(obj, nits = 15, couch = F, startE = F, varind = 0.2,
  uprat = 0.1, GModel = "New Lester", quiet = F)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{nits}{integer value, the number of iterations for resolving the IFD}

\item{couch}{logical, has the couch effect been calculated and present in slot 'couch' of the landscape object.}

\item{startE}{logical, is the IFD calculation starting at previous effort vector in slot 'eff' of the landscape object (allows further IFD iterations)}

\item{varind}{positive real number, the slope of the logit choice (low values are 'choosey' a value of 1 is proportional to utility)}

\item{uprat}{fraction, When converging on a stable effort distribution this is the contributing fraction of new effort}

\item{quiet}{logical, should warnings and messages be surpressed?}

\item{Gmodel}{character, the density dependent growth model. Options are 'Lester1', 'Lester2' and 'New Lester'}
}
\description{
The IFD effort calculator
}
\author{
T. Carruthers
}
