% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{DelMan}
\alias{DelMan}
\title{Delete a management option from a Landscape object}
\usage{
DelMan(obj, todel = 1)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{tocopy}{positive integer greater than 1, the number of replicate management options}
}
\description{
Delete a management option from a Landscape object
}
\author{
T. Carruthers
}
