% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{Ecomp}
\alias{Ecomp}
\title{Effort comparison observed (experimental) vs predicted by the IFD calculation}
\usage{
Ecomp(obj, funcy = NA, lcex = 0.7, pcol = NA, pcex = 0.7,
  log = T, nice = T, lims = NA, forpaper = F, dens = F,
  plottext = T, Rstat = T)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{funcy}{real number, an optional function minimization value that can be plotted}

\item{lcex}{positive real number, the label font size}

\item{pcol}{character, the color of the points}

\item{pcex}{positive real number, the size of the points}

\item{log}{logical, should log effort be plotted?}

\item{nice}{logical, should a word cloud arrangement of lake labels be used (can be unstable)?}

\item{lims}{vector 2 long, x and y axis limit}

\item{forpaper}{logical, journal style changes}

\item{dens}{logical, should effort density be plotted?}

\item{plottex}{logical, should lake names be added?}

\item{Rsta}{logical, should the coefficient of determination ('R-squared') value be plotted?}
}
\description{
Effort comparison observed (experimental) vs predicted by the IFD calculation
}
\author{
T. Carruthers
}
