% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{esimlakemanage}
\alias{esimlakemanage}
\title{Create a dummy set of lake attributes for initializing the landscape object}
\usage{
esimlakemanage(attrtype, nl, exn)
}
\arguments{
\item{attrtype}{a vector of attribute types}

\item{nl}{integer a number of lakes}

\item{exn}{integer the number of alternative management options under investigation}
}
\description{
Create a dummy set of lake attributes for initializing the landscape object
}
\author{
T. Carruthers
}
