% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{fac2num}
\alias{fac2num}
\title{Convert factors to numeric values}
\usage{
fac2num(x)
}
\arguments{
\item{x}{factor}
}
\description{
Convert factors to numeric values
}
\examples{
\dontrun{
  fac2num(as.factor(1))
}
}
\author{
T. Carruthers
}
