% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{get_manage_slots}
\alias{get_manage_slots}
\title{Returns the names of slots that have a dimension 'management'}
\usage{
get_manage_slots(obj)
}
\arguments{
\item{obj}{an object of class 'Landscape' that has only 1 management policy (nmanage=1)}
}
\description{
Returns the names of slots that have a dimension 'management'
}
\author{
T. Carruthers
}
