% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{get_sim_slots}
\alias{get_sim_slots}
\title{Returns the names of slots that have a dimension 'simulation'}
\usage{
get_sim_slots(obj)
}
\arguments{
\item{obj}{an object of class 'Landscape' that has only 1 simulation (nsim=1)}
}
\description{
Returns the names of slots that have a dimension 'simulation'
}
\author{
T. Carruthers
}
