% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{getattr}
\alias{getattr}
\title{This is the look up table for the values of lake attributes}
\usage{
getattr(vals, dims)
}
\arguments{
\item{vals}{a vector of attribute value}

\item{dims}{the number of levels of each attribute}
}
\description{
This is the look up table for the values of lake attributes
}
\author{
T. Carruthers
}
