% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{getbias}
\alias{getbias}
\title{Simulate lognormally sampled biases based on a CV}
\usage{
getbias(nsim, CV)
}
\arguments{
\item{nsim}{the number of simulations}

\item{CV}{the coefficient of variation of the distribution of biases}
}
\description{
Simulate lognormally sampled biases based on a CV
}
\examples{
\dontrun{
plot(density(getbias(nsim=1000,CV=0.1)))
}
}
\author{
T. Carruthers
}
