% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{getEdist}
\alias{getEdist}
\title{This is just a way of initializing the distance matrix between pop centres and lakes as the Euclidean distance.}
\usage{
getEdist(xfrom, yfrom, xto, yto, nmanage, CVd, stand)
}
\arguments{
\item{xfrom}{a vector of longitudes of some points e.g. pop centres}

\item{yfrom}{a vector of latitudes of some points e.g. pop centres}

\item{xto}{a vector of longitude of some other points e.g. lakes}

\item{yto}{a vector of latitudes of some other points e.g. lakes}

\item{nmanage}{the number of alternative management options under investigation}

\item{CVd}{the CV on lognormal distances (for added stochasticity)}

\item{stand}{Boolean: should all distances be standardized to mean 1}
}
\description{
This is just a way of initializing the distance matrix between pop centres and lakes as the Euclidean distance.
}
\author{
T. Carruthers
}
