% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{geterr}
\alias{geterr}
\title{Generates lognormal random variable}
\usage{
geterr(n, CV)
}
\arguments{
\item{n}{integer: the number of draws}

\item{CV}{numeric: the coefficient of variation}
}
\description{
Generates lognormal random variable
}
\author{
T. Carruthers
}
