% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{getlxslev}
\alias{getlxslev}
\title{Initializes a stocking array, nmanage x lake x stocking type.}
\usage{
getlxslev(areas, nst, nmanage, stwt)
}
\arguments{
\item{areas}{a list of named stocked resources (lakes)}

\item{nst}{integer: the number of stocking types on the landscape}

\item{nmanage}{integer: the number of alternative management options}

\item{stwt}{numeric the weight of the stocked fish of each stocking type}
}
\description{
Initializes a stocking array, nmanage x lake x stocking type.
}
\author{
T. Carruthers
}
