% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_stochasticity.R
\name{getmvlogerr}
\alias{getmvlogerr}
\title{Generate multivariate logistic error}
\usage{
getmvlogerr(x, CV, nsim)
}
\arguments{
\item{x}{vector of positive real numbers}

\item{CV}{real positive number, the coefficient of variation}

\item{nsim}{positive integer, the number of simulations}
}
\description{
Generate multivariate logistic error
}
\author{
T. Carruthers
}
