% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_object.R
\docType{methods}
\name{initialize,Landscape-method}
\alias{initialize,Landscape-method}
\title{Initialization of a landscape object}
\usage{
\S4method{initialize}{Landscape}(.Object, templatefile = "bla")
}
\arguments{
\item{templatefile}{character the location of a landscape .csv template file}

\item{object}{object of class MSE}
}
\description{
Initialization of a landscape object
}
