% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_object.R
\docType{class}
\name{Landscape-class}
\alias{Landscape-class}
\title{Class \code{'Landscape'}}
\description{
An S4 class object for storing all attributes of the landscape including anglers, population centres and lakes.
Landscape objects can be modified, subsetted and adapted using a range of SSES functions.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character string. The name of the Data object. Single value. Character string}

\item{\code{Note}}{Character string. Any relevant information}

\item{\code{npc}}{Integer. The number of population centres}

\item{\code{nl}}{Integer. The number of lakes}

\item{\code{na}}{Integer. The number of angler classes}

\item{\code{nattr}}{Integer. The number of lake attributes.}

\item{\code{ncat}}{Integer. The number of categorical lake attributes.}

\item{\code{nsim}}{Integer. The number of simulations (landscape simulations per management mode).}

\item{\code{nmanage}}{Integer. The number of management modes.}

\item{\code{nage}}{Integer. The maximum number of fish age classes modelled}

\item{\code{nst}}{Integer. The number of stocking types (e.g. fry, fingerlings etc)}

\item{\code{nattrvals}}{Integer. The maximum number of attribute values for each categorical lake attribute}

\item{\code{pcnam}}{Character value. The name of each population centre (population centre).}

\item{\code{pcsize}}{Array. The number of anglers in each population (sim, population centre).}

\item{\code{pcx}}{Numeric. The longitude of each population centre (population centre).}

\item{\code{pcy}}{Numeric. The latitude of each population centre (population centre).}

\item{\code{lakenam}}{Character. The name of each lake (short form lake code) (lake).}

\item{\code{longnam}}{Character. The long name of each lake (lake).}

\item{\code{lakex}}{Numeric. The longitude of each lake (lake).}

\item{\code{lakey}}{Numeric. The latitute of each lake (lake).}

\item{\code{lakearea}}{Numeric. The surface area of the lake in hectares (lake).}

\item{\code{lxslev}}{3D Array. The stocking level (management mode, lake, stocking type).}

\item{\code{GDD}}{Array. Growing degree days for each lake (sim, lake).}

\item{\code{TDS}}{Array Total disolved solids for each lake (sim, lake).}

\item{\code{stockable}}{NOT IN USE. Whether the lake is stocked (lake).}

\item{\code{anam}}{Character. The name of the angler classes (angler class).}

\item{\code{apr}}{3D Array. The angler participation rate - mean effort per license (simulation, population centre, angler class).}

\item{\code{aq}}{Array. The angler catchabilty coefficient (sim, angler class).}

\item{\code{attr}}{data.frame. The attributes of each angler class.}

\item{\code{pcxl}}{3D array. The distance of each population centre to each lake, pc x l (management mode, population centre, lake).}

\item{\code{pcxa}}{3D array. The fraction of each angler class in each population centre, pc x a (simulation, population centre, angler class).}

\item{\code{lxattr}}{3D array. Lake attributes l x attr (management mode, lake, lake attribute).}

\item{\code{axattr}}{3D array. Angler attributes a x attr (sim, population centre, angler class).}

\item{\code{eff}}{5D array. Output. The calculated effort on each lake (management option, simulation, population centre, lake, angler class).}

\item{\code{avs}}{Array. Output. Average size of fish caught on each lake.}

\item{\code{cr}}{Array. Output. Catch rate on each lake.}

\item{\code{exeff}}{Array. Input. Experimental observed effort.}

\item{\code{exavs}}{Array. Input. Experimental average size of fish caught.}

\item{\code{excr}}{Array. Input. Experimental observed catch rate.}

\item{\code{sigeff}}{3D array. Effort summed (management mode, simulation, lake).}

\item{\code{sigexeff}}{3D array. Observed experimental effort summed (management mode, simulation, lake).}

\item{\code{cont}}{List. List of polygons that are map contours or other map features to be plotted.}

\item{\code{topcols}}{Character. A vector of colors matching each item in the list cont.}

\item{\code{costs}}{4D Array. Management costs (simulation, lake, attribute, attribute level).}

\item{\code{Scosts}}{3D Array. Stocking costs per fish (simulation, lake, stocking type).}

\item{\code{totcost}}{3D Array. Total current stocking costs per lake (management, simulation, lake).}

\item{\code{effval}}{2D Array. The current value of a unit of effort (simululation, angler class).}

\item{\code{licval}}{2D Arrayl. The value of a license sale (simulation, 1).}

\item{\code{stnam}}{Character. The name of each stocking type (stocking type).}

\item{\code{aclass}}{Integer. The age class of each stocking type (stocking type).}

\item{\code{stwt}}{Numeric. The weight (gms) that each stocking type is stocked at (stocking type).}

\item{\code{stlen}}{Numeric. The lenght (mm) that each stocking type is stocked at (stocking type).}

\item{\code{popval}}{Array. Population dynamics parameters (sim, population parameter, stocking type).}

\item{\code{poperr}}{Array. The error structure and variance for each population parameter (population parameter, (type, CV))}

\item{\code{errs}}{Data.frame. A set of errors (log normal CVs) for various other attribtues of the model such as GDD and population size}

\item{\code{Mage}}{3D array. Mortality at age (simulation, age class, stocking type)}

\item{\code{CB}}{2D array. Output. Total Benefit / Cost. Total effort per $ stocking (management mode, simulation)}

\item{\code{conv}}{2D array. Output. The total effort convergence information (lake, iteration of algorithm)}

\item{\code{Econv}}{2D array. Output. The effort convergence information (lake, iteration of algorithm)}

\item{\code{lm}}{Hierarchical list. R linear models for the continuous variables (continuous variable)(angler class)}

\item{\code{U}}{Array. A record of utility for the first management mode and simulation (population centre, angler class)}

\item{\code{FTroutAng}}{Numeric. The fraction of licensed anglers that are trout fishers (population centre)}

\item{\code{acc}}{Array. Access metric (simulation, lake)}

\item{\code{acc_a}}{Numeric. The optimized slope parameter converting the access metric into a lake gravity term}

\item{\code{misc}}{list. Somewhere to store stuff}

\item{\code{sel}}{Array. Age selectivity of anglers (simulation, age)}

\item{\code{GModel}}{Character. Growth model used (Lester1, Lester2, New Lester)}

\item{\code{couch}}{4D Array. A specified couch effect (managment mode, simulation, population centre, angler class)}

\item{\code{fac}}{DISUSED}

\item{\code{fac_a}}{DISUSED}

\item{\code{DR}}{Numeric. Discard rate. The fraction of fish that are discarded.}

\item{\code{DD}}{Numeric. Dead Discarding. The fraction of fish that are discarded that die (simulation)}

\item{\code{BagLim}}{Array. Bag limit. The maximum number of fish that may be retained per day on each lake  (management mode, lake)}

\item{\code{maxdays}}{Integer. Maximum days per licence if doing calculations with couch effect = TRUE}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('Landscape', datafile)}
}

\examples{

myland<-new('Landscape',datafile)

}
\author{
T. Carruthers
}
\keyword{classes}
