% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{Lsub}
\alias{Lsub}
\title{Subset the landscape object by lake}
\usage{
Lsub(obj, ind)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{ind}{integer vector: the lakes to subset by in the position they are listed in the lakenam slot}
}
\description{
Subset the landscape object by lake
}
\author{
T. Carruthers
}
