% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make a standard color transparent}
\usage{
makeTransparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{character string, a color}

\item{alpha}{percentage, the transparency of the color}
}
\description{
Make a standard color transparent
}
\author{
T. Carruthers
}
