% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{mapeff}
\alias{mapeff}
\title{Plot a map of landscape effort}
\usage{
mapeff(obj, maxps = 2, latlim = NA, lonlim = NA, nman = 1,
  nsim = 1, dens = F, U = F, lcex = 0.55, pcex = 0.75,
  lpick = NA)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{maxps}{real number, maximim size of points}

\item{latlim}{vector 2 long, the x axis, latitude range of the plot}

\item{lonlim}{vector 2 long, the y axis, longitude range of the plot}

\item{nman}{positive integer, the number of managmeent options to plot}

\item{nsim}{positive integer, the number of simulations to plot}

\item{dens}{logical, should effort density be plotted?}

\item{U}{logical, should utility be plotted instead of effort?}

\item{lcex}{the label font size}

\item{pcex}{the point size}

\item{lpick}{vector of either lake names or their index number, to be highlighted}

\item{nice}{logical, should a word cloud arrangement of lake labels be used (can be unstable)?}
}
\description{
Plot a map of landscape effort
}
\author{
T. Carruthers
}
