% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{ninety}
\alias{ninety}
\title{Get 10th and 90th percentiles of a vector of values}
\usage{
ninety(x)
}
\arguments{
\item{x}{a vector of values numeric values}
}
\description{
Get 10th and 90th percentiles of a vector of values
}
\examples{
\dontrun{
ninety(c(NA,1,NA,2,NA))
}
}
\author{
T. Carruthers
}
