% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{nona}
\alias{nona}
\title{Are there no NA values in a vector?}
\usage{
nona(x)
}
\arguments{
\item{x}{a vector of values numeric, or character}
}
\description{
Are there no NA values in a vector?
}
\examples{
\dontrun{
nona(c(4,1,NA,2,6))
}
}
\author{
T. Carruthers
}
