% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_fitting.R
\name{opt_DR_acc}
\alias{opt_DR_acc}
\title{Optimize for slope on accessibility values}
\usage{
opt_DR_acc(
  par,
  obj,
  nits = 15,
  opt = T,
  ploty = F,
  varind = 0.15,
  uprat = 0.05,
  log = T,
  quiet = T
)
}
\arguments{
\item{par}{A single parameter estimating the relative impact of the accessibility on overall utility}

\item{obj}{an object of class 'Landscape'}

\item{nits}{integer value, the number of iterations for resolving the IFD}

\item{opt}{logical, return funciton value or object?}

\item{ploty}{logical, plot the fit as the optimizer runs?}

\item{varind}{positive real number, the slope of the logit choice (low values are 'choosey' a value of 1 is proportional to utility)}

\item{uprat}{fraction, When converging on a stable effort distribution this is the contributing fraction of new effort}

\item{log}{logical, plot log effort?}

\item{quiet}{logical, suppress outputs while fitting?}
}
\description{
Optimize for slope on accessibility values
}
\author{
T. Carruthers
}
