% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{pc_simp}
\alias{pc_simp}
\title{Landscape simplification (to ensure identifiability in IFD effort - a distinct lake x population centre distance matrix)}
\usage{
pc_simp(obj, h = 1000)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{h}{postive real number, the height of the dendrogram for cutting (population centre simplification)}
}
\description{
For all landscape characteristics that are uncertain, these parameters are sampled.
}
\author{
T. Carruthers
}
