% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{plotconv}
\alias{plotconv}
\title{Plot IFD convergence diagnostics}
\usage{
plotconv(obj, tol = 0.1, totE = F, n = 20, ylimy = NA,
  forpaper = F)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{tol}{positive real number, the tolerance of the detection of convergence (absolute effort difference}

\item{totE}{logical, should the convergence diagnostic run on total lake effort (as opposed to angler x lake effort)?}

\item{n}{positive integer, the number of convergence data points to plot}

\item{ylimy}{vector length 2, the y axis limits}

\item{forpaper}{logical, journal style changes}
}
\description{
Plot IFD convergence diagnostics
}
\author{
T. Carruthers
}
