% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{plotconv2}
\alias{plotconv2}
\title{Plot IFD convergence diagnostics version 2}
\usage{
plotconv2(obj, totE = F, n = 20, ylimy = NA, forpaper = F)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{totE}{logical, should the convergence diagnostic run on total lake effort (as opposed to angler x lake effort)?}

\item{n}{positive integer, the number of convergence data points to plot}

\item{ylimy}{vector length 2, the y axis limits}

\item{forpaper}{logical, journal style changes}
}
\description{
Plot IFD convergence diagnostics version 2
}
\author{
T. Carruthers
}
