% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{replic8}
\alias{replic8}
\title{Expand a Lansdscape object to mulitple management regimes}
\usage{
replic8(obj, nmanage = NA, nsim = NA)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{nmanage}{positive integer greater than 1, the number of replicate management options}

\item{nsim}{positive integer greater than 1, the number of replicate simulations}
}
\description{
Expand a Lansdscape object to mulitple management regimes
}
\author{
T. Carruthers
}
