% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_stochasticity.R, R/Misc_exported.R
\name{sdconv}
\alias{sdconv}
\title{Convert a desired mean and standard deviation into a log normal sd}
\usage{
sdconv(m, sd)

sdconv(m, sd)
}
\arguments{
\item{m}{real number, mean value}

\item{sd}{real positive number, sd of the variable}

\item{m}{mean value of transformed variable}

\item{sd}{standard deviation of transformed variable}
}
\description{
Convert a desired mean and standard deviation into a log normal sd

Get lognormal standard deviation parameter for a variable with m and sd transformed values
}
\examples{
\dontrun{
sdconv(m=1,sd=0.5)
}
}
\author{
T. Carruthers

T. Carruthers
}
