% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSES-package.R
\docType{package}
\name{SSES-package}
\alias{SSES-package}
\alias{SES-package}
\alias{DLMtool}
\title{Spatial Social Ecological Systems}
\description{
Strategic landscape-scale management of BC recreational trout lake fisheries using advanced SES modelling
}
\references{
Carruthers, T. R., K. Dabrowska, W. Haider, E. A. Parkinson, D. A. Varkey, H. Ward,
M. K. McAllister, T. Godin, Van Poorten, B., Askey, P.J., Wilson, K.L., Hunt, L.M., Clarke A.,
Newton, E., Walters, C., Post J.R. Landscape scale social and ecological outcomes of dynamic
angler and fish behaviours: Processes, data, and patterns.,
Can. J. Fish. Aquat. Sci. (2018). https://doi.org/10.1139/cjfas-2018-0168
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{angling,}
\keyword{ecological}
\keyword{evaluation}
\keyword{fisheries,}
\keyword{management}
\keyword{recreational}
\keyword{social}
\keyword{strategy}
\keyword{systems,}
\keyword{trout,}
