% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{std}
\alias{std}
\title{Standardize an effort convergence time series, relative to the last observation}
\usage{
std(x)
}
\arguments{
\item{x}{vector of effort values, nits long}
}
\description{
Standardize an effort convergence time series, relative to the last observation
}
\author{
T. Carruthers
}
