% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R
\name{Stoch}
\alias{Stoch}
\title{Add stochasticity to the landscape object}
\usage{
Stoch(obj, nsim)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{nsim}{integer: the number of simulations (stochastic samples of parameters)}
}
\description{
For all landscape characteristics that are uncertain, these parameters are sampled.
}
\author{
T. Carruthers
}
