% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_object.R
\docType{methods}
\name{summary,Landscape-method}
\alias{summary,Landscape-method}
\title{Summary of a landscape object}
\usage{
\S4method{summary}{Landscape}(object, lcex = 0.4, pcex = 0.8,
  maxps = 4)
}
\arguments{
\item{object}{object of class MSE}
}
\description{
Summary of a landscape object
}
