% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{tna}
\alias{tna}
\title{Count NA values in a vector}
\usage{
tna(x)
}
\arguments{
\item{x}{a vector of values numeric, or character}
}
\description{
Count NA values in a vector
}
\examples{
\dontrun{
tna(c(NA,1,NA,2,NA))
}
}
\author{
T. Carruthers
}
