% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{ttp}
\alias{ttp}
\title{Text positioning for word cloud}
\usage{
ttp(
  x,
  y,
  words,
  cex = 1,
  pcex = 1,
  pcol = "red",
  new = TRUE,
  show.lines = TRUE,
  ...
)
}
\arguments{
\item{x}{vector x coordinates}

\item{y}{vector of y coordinates}

\item{cex}{font size}

\item{pcex}{point size}

\item{pcol}{color of points}

\item{new}{logical, new plot needed?}

\item{show.lines}{logical, should lines between points and text be included}
}
\description{
Text positioning for word cloud
}
\author{
T. Carruthers
}
