% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_plotting.R
\name{ttp_pos}
\alias{ttp_pos}
\title{Text positioning for word cloud (just data points)}
\usage{
ttp_pos(
  x,
  y,
  words,
  cex = 1,
  pcex = 1,
  pcol = "red",
  new = TRUE,
  show.lines = TRUE,
  lwd = 1,
  lcol = "black",
  ...
)
}
\arguments{
\item{x}{vector x coordinates}

\item{y}{vector of y coordinates}

\item{cex}{font size}

\item{pcex}{point size}

\item{pcol}{color of points}

\item{new}{logical, new plot needed?}

\item{show.lines}{logical, should lines between points and text be included}

\item{lwd}{postive real number, the width of the lines}

\item{lcol}{character, the color of the lines}

\item{works}{character vector of words to plot}
}
\description{
Text positioning for word cloud (just data points)
}
\author{
T. Carruthers
}
