% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_stochasticity.R, R/Misc_exported.R
\name{geterr}
\alias{geterr}
\title{Generate a log normal error}
\usage{
geterr(n, CV)

geterr(n, CV)
}
\arguments{
\item{n}{integer: the number of draws}

\item{CV}{numeric: the coefficient of variation}
}
\description{
Generate a log normal error

Generates lognormal random variable
}
\author{
T. Carruthers

T. Carruthers
}
