% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_exported.R
\name{sdconv}
\alias{sdconv}
\title{Get lognormal standard deviation parameter for a variable with m and sd transformed values}
\usage{
sdconv(m, sd)
}
\arguments{
\item{m}{mean value of transformed variable}

\item{sd}{standard deviation of transformed variable}
}
\description{
Get lognormal standard deviation parameter for a variable with m and sd transformed values
}
\examples{
\dontrun{
sdconv(m=1,sd=0.5)
}
}
\author{
T. Carruthers
}
