% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape_manipulation_functions.R,
%   R/Landscape_stochasticity.R
\name{Stoch}
\alias{Stoch}
\title{Add stochasticity to the landscape object}
\usage{
Stoch(obj, nsim)

Stoch(obj, nsim)
}
\arguments{
\item{obj}{an object of class 'Landscape'}

\item{nsim}{integer: the number of simulations (stochastic samples of parameters)}

\item{obj}{an object of class 'Landscape'}

\item{nsim}{integer value, the number of simulated landscapes}
}
\description{
For all landscape characteristics that are uncertain, these parameters are sampled.

this function uses the landscape slot 'errs' to generate stochasticity in landscape variables according to
a coefficient of variation. For most variables this is a lognormal error term.
}
\author{
T. Carruthers

T. Carruthers
}
